move default global-error into a separate file
include global error chunk
