+++ packages/next/src/client/components/app-router.tsx
@@ -23,11 +23,10 @@ import {
   PathParamsContext,
 } from '../../shared/lib/hooks-client-context.shared-runtime'
 import { dispatchAppRouterAction, useActionQueue } from './use-action-queue'
-import {
-  default as DefaultGlobalError,
-  ErrorBoundary,
+import { ErrorBoundary } from './error-boundary'
+import DefaultGlobalError, {
   type GlobalErrorComponent,
-} from './error-boundary'
+} from '../../client/components/global-error'
 import { isBot } from '../../shared/lib/router/utils/is-bot'
 import { addBasePath } from '../add-base-path'
 import { AppRouterAnnouncer } from './app-router-announcer'
