+++ packages/next/src/server/app-render/entry-base.ts
@@ -29,8 +29,11 @@ import * as serverHooks from '../../client/components/hooks-server-context'
 import { HTTPAccessFallbackBoundary } from '../../client/components/http-access-fallback/error-boundary'
 import { createMetadataComponents } from '../../lib/metadata/metadata'
 import { patchFetch as _patchFetch } from '../lib/patch-fetch'
-// not being used but needs to be included in the client manifest for /_not-found
+// Not being directly used but should be included in the client manifest for /_not-found
+// * ErrorBoundary -> client/components/error-boundary
+// * GlobalError -> client/components/global-error
 import '../../client/components/error-boundary'
+import '../../client/components/global-error'
 import {
   MetadataBoundary,
   ViewportBoundary,
